criação de equipe exception

+++ src/main/java/org/serratec/java2backend/alterdatapac6/service/EquipeService.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import org.serratec.java2backend.alterdatapac6.dto.EquipeDto;
 import org.serratec.java2backend.alterdatapac6.entity.EquipeEntity;
+import org.serratec.java2backend.alterdatapac6.exceptions.EquipeDuplicadaException;
+import org.serratec.java2backend.alterdatapac6.exceptions.EquipeNotFoundException;
 import org.serratec.java2backend.alterdatapac6.mapper.EquipeMapper;
 import org.serratec.java2backend.alterdatapac6.repository.EquipeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +13,7 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class EquipeService {
-	
+
 	@Autowired
 	EquipeRepository repository;
 
@@ -23,31 +25,49 @@ public class EquipeService {
 		return repository.findAll();
 	}
 
-	public EquipeEntity getByNome(String nome) {
-		return repository.getByNome(nome);
+	public EquipeEntity getByNome(String nome) throws EquipeNotFoundException {
+		
+		EquipeEntity equipe= repository.getByNome(nome);
+		
+		if(equipe==null) {
+			throw new EquipeNotFoundException("Equipe "+nome+" não encrontrada!");
+		}
+		
+		return equipe;
 	}
 
-	public EquipeDto create(EquipeDto equipe) {
+	public EquipeDto create(EquipeDto equipe) throws EquipeDuplicadaException {
+		EquipeEntity equipeBd = repository.getByNome(equipe.getNome());
+
+		if (equipeBd != null) {
+			throw new EquipeDuplicadaException("A equipe " + equipe.getNome() + " já exite! Escolha outro nome.");
+		}
+
 		EquipeEntity equipeNova = mapper.toEntity(equipe);
 		return mapper.toDto(repository.save(equipeNova));
 	}
 
-	public EquipeDto update(String nomeEquipe, EquipeDto equipe) {
+	public EquipeDto update(String nomeEquipe, EquipeDto equipe) throws EquipeDuplicadaException {
 		EquipeEntity equipeNova = mapper.toEntity(equipe);
 		EquipeEntity equipeHist = repository.getByNome(nomeEquipe);
 
-		if(equipeNova.getNome()!=null) {
+		if (equipeNova.getNome() != null) {
+
+			EquipeEntity equipeBd = repository.getByNome(equipe.getNome());
+
+			if (equipeBd != null) {
+				throw new EquipeDuplicadaException("A equipe " + equipe.getNome() + " já exite! Escolha outro nome.");
+			}
 			equipeHist.setNome(equipeNova.getNome());
 		}
-		
+
 		if (equipeNova.getDescricao() != null) {
 			equipeHist.setDescricao(equipeNova.getDescricao());
 		}
 		return mapper.toDto(repository.save(equipeHist));
 
 	}
-	
-	
+
 	/*
 	 * public EquipeDto update(EquipeDto equipe) { EquipeEntity equipeNova =
 	 * mapper.toEntity(equipe); EquipeEntity equipeHist =
@@ -62,12 +82,14 @@ public class EquipeService {
 	 * }
 	 */
 
-	public void deleteByNome(String nome) {
+	public void deleteByNome(String nome) throws EquipeNotFoundException {
 		EquipeEntity equipe = getByNome(nome);
 		Long equipeId = equipe.getId();
-
+		
+		if(equipe==null) {
+			throw new EquipeNotFoundException("Equipe "+nome+" não encrontrada!");
+		}
 		repository.deleteById(equipeId);
 	}
-
 
 }
